refactor(engine): split stat source specs
